fixed warning on ColdFire

+++ Drivers/freeRTOS/tasks.c
@@ -3576,7 +3576,7 @@ TCB_t *pxTCB;
 
 		/* Pad the end of the string with spaces to ensure columns line up when
 		printed out. */
-		for( x = strlen( pcBuffer ); x < ( configMAX_TASK_NAME_LEN - 1 ); x++ )
+		for( x = (long)strlen( pcBuffer ); x < ( configMAX_TASK_NAME_LEN - 1 ); x++ )
 		{
 			pcBuffer[ x ] = ' ';
 		}
